refactor: revert test files movement back to test/lints
